--- plugins/com.google.cloud.tools.eclipse.dataflow.ui.test/src/com/google/cloud/tools/eclipse/dataflow/ui/page/MessagesTest.java
@@ -1,44 +1,0 @@
-/*
- * Copyright 2016 Google Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.cloud.tools.eclipse.dataflow.ui.page;
-
-import org.junit.Assert;
-import org.junit.Test;
-
-import com.google.cloud.tools.eclipse.dataflow.ui.Messages;
-
-public class MessagesTest {
-
-  @Test
-  public void testRunOptions() {
-    Assert.assertEquals(
-        "Set Default Cloud Dataflow Run Options", Messages.getString("SET_RUN_OPTIONS"));
-  }
-  
-  @Test
-  public void testWizardDescription() {
-    Assert.assertEquals(
-        "This wizard creates a new Google Cloud Dataflow project.",
-        Messages.getString("WIZARD_DESCRIPTION"));
-  }
-  
-  @Test
-  public void testLocationToolTip() {
-    Assert.assertEquals("The location where the project will be created. Must be an existing local directory.",
-        Messages.getString("LOCATION_TOOLTIP"));
-  }
-}
